change type to size_t

+++ include/webcachesim/filters/bloom.h
@@ -747,7 +747,7 @@ public:
 //        uint64_t bits, var = (age < 0) ? -age : age;
 //        for (bits = 0; var != 0; ++bits) var >>= 1;
         float bucket_base = log2(age);
-        uint bucket = bucket_base * bucket_detail;
+        size_t bucket = bucket_base * bucket_detail;
         int i = min(bucket, obj_freq_buckets.size() - 1);
 
         int obj_j = min(obj_count - 1, freq_bucket_count - 1);
